// src/runtime/types/common.d.ts
export type AlignmentType = 'left' | 'center' | 'right'
export type SeoTagType = 'div' | 'h1' | 'h2' | 'h3'
export type BackgroundAttachmentType = 'fixed' | 'local' | 'scroll'
export type HorizontalAlignType = 'left' | 'center' | 'right'
export type VerticalAlignType = 'top' | 'middle' | 'bottom'
export type BackgroundPositionType =
    | 'top left'
    | 'top center'
    | 'top right'
    | 'center left'
    | 'center center'
    | 'center right'
    | 'bottom left'
    | 'bottom center'
    | 'bottom right'